+++ lib/flutter_sfs.dart
@@ -3,4 +3,3 @@ library flutter_sfs;
 export 'src/sfs/sfs_init.dart';
 export 'src/sfs/sfs.dart';
 export 'src/utils.dart';
-export 'src/navigator_service.dart';